refactor: move Prometheus response types to ui-core/types/metrics
Moves PrometheusQueryRange* types from model-serving/shared/types to
ui-core/types/metrics where they belong alongside other metrics
infrastructure types. Removes false gpuaas → model-serving dependency
that existed solely for these types.
