+++ frontend/src/__mocks__/mockPrometheusBias.ts
@@ -2,7 +2,7 @@
 import type {
   PrometheusQueryRangeResponse,
   PrometheusQueryRangeResponseDataResult,
-} from '@odh-dashboard/model-serving/shared/types';
+} from '@odh-dashboard/ui-core/types/metrics';
 
 type MockPrometheusServingType = {
   result?: PrometheusQueryRangeResponseDataResult[];
