+++ frontend/src/__mocks__/mockPrometheusServing.ts
@@ -1,7 +1,7 @@
 import type {
   PrometheusQueryRangeResponse,
   PrometheusQueryRangeResponseDataResult,
-} from '@odh-dashboard/model-serving/shared/types';
+} from '@odh-dashboard/ui-core/types/metrics';
 
 type MockPrometheusServingType = {
   result?: PrometheusQueryRangeResponseDataResult[];
