+++ frontend/src/api/prometheus/__tests__/usePrometheusQueryRange.spec.ts
@@ -1,6 +1,6 @@
 import { act } from 'react';
 import { testHook } from '@odh-dashboard/jest-config/hooks';
-import type { PrometheusQueryRangeResponseData } from '@odh-dashboard/model-serving/shared/types';
+import type { PrometheusQueryRangeResponseData } from '@odh-dashboard/ui-core/types/metrics';
 import axios from '#~/utilities/axios';
 import { mockPrometheusServing } from '#~/__mocks__/mockPrometheusServing';
 import usePrometheusQueryRange from '#~/api/prometheus/usePrometheusQueryRange';
