+++ frontend/src/api/prometheus/serving.ts
@@ -1,13 +1,14 @@
 import * as React from 'react';
 import { SupportedArea, useIsAreaAvailable } from '@odh-dashboard/plugin-core/areas';
 import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
-import { RefreshIntervalTitle, TimeframeTitle } from '@odh-dashboard/ui-core/types/metrics';
-import { RefreshIntervalValue } from '@odh-dashboard/ui-core/utilities/metrics';
 import {
-  PerformanceMetricType,
-  PrometheusQueryRangeResponseDataResult,
-  PrometheusQueryRangeResultValue,
-} from '@odh-dashboard/model-serving/shared/types';
+  RefreshIntervalTitle,
+  TimeframeTitle,
+  type PrometheusQueryRangeResponseDataResult,
+  type PrometheusQueryRangeResultValue,
+} from '@odh-dashboard/ui-core/types/metrics';
+import { RefreshIntervalValue } from '@odh-dashboard/ui-core/utilities/metrics';
+import { PerformanceMetricType } from '@odh-dashboard/model-serving/shared/types';
 import { ModelMetricType } from '#~/pages/modelServing/screens/metrics/ModelServingMetricsContext';
 import useRefreshInterval from '#~/utilities/useRefreshInterval';
 import { PROMETHEUS_BIAS_PATH } from '#~/api/prometheus/const';
