+++ packages/model-serving/src/components/metrics/types.ts
@@ -2,7 +2,7 @@ import { DomainTuple, ForAxes } from 'victory-core';
 import { ChartThemeDefinitionInterface } from '@patternfly/react-charts/victory';
 import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
 import type { BiasMetricType, BiasMetricConfig } from '@odh-dashboard/trustyai/types';
-import type { PrometheusQueryRangeResultValue } from '@odh-dashboard/model-serving/shared/types';
+import type { PrometheusQueryRangeResultValue } from '@odh-dashboard/ui-core/types/metrics';
 import { ModelMetricType } from './ModelServingMetricsContext';
 
 export type TranslatePoint = (line: GraphMetricPoint) => GraphMetricPoint;
