+++ packages/ui-core/src/types/metrics.ts
@@ -33,3 +33,23 @@ export type RefreshIntervalValueType = {
 export type BreadcrumbItemType = {
   label: string;
 } & EitherNotBoth<{ link: string }, { isActive: boolean }>;
+
+export type PrometheusQueryRangeResultValue = [number, string];
+
+export type PrometheusQueryRangeResponseDataResult = {
+  metric: {
+    request?: string;
+    pod?: string;
+  };
+  values: PrometheusQueryRangeResultValue[];
+};
+
+export type PrometheusQueryRangeResponseData = {
+  result?: PrometheusQueryRangeResponseDataResult[];
+  resultType: string;
+};
+
+export type PrometheusQueryRangeResponse = {
+  data: PrometheusQueryRangeResponseData;
+  status: string;
+};
